# Conflicts:
#	app/src/main/java/com/highway/common/base/activity/DashBoardActivity.java
#	app/src/main/java/com/highway/drivermodule/driverFragment/InvoiceBottomDialogFragment.java
#	app/src/main/java/com/highway/drivermodule/driverFragment/RatingBottomDialogFragment.java

+++ app/src/main/java/com/highway/customer/customerActivity/BookingWithDetailsActivity.java
@@ -89,7 +89,7 @@ public class BookingWithDetailsActivity extends AppCompatActivity implements OnM
     public TextView bookTruckTv, phoneNoTv, nameTv;
     public String userName, userMobNo;
     Button getDirection;
-    float distance;
+    int distance;
     MarkerOptions markerOptions1;
     MarkerOptions markerOptions2;
     List<Marker> markers = new ArrayList<>();
@@ -289,7 +289,7 @@ public class BookingWithDetailsActivity extends AppCompatActivity implements OnM
         Location endPoint = new Location("destName");
         endPoint.setLatitude(destLatitude);
         endPoint.setLongitude(destLongitude);
-        distance =  startPoint.distanceTo(endPoint) / 1000;
+        distance = (int) (startPoint.distanceTo(endPoint) / 1000);
 
         Intent intent = new Intent();
         intent.putExtra("distance", distance);
